// src/enums/Skill.ts
export enum Skill {
  Astral = "astral",
  Athletics = "athletics",
  Biotech = "biotech",
  CloseCombat = "closeCombat",
  Con = "con",
  Conjuring = "conjuring",
  Cracking = "cracking",
  Electronics = "electronics",
  Enchanting = "enchanting",
  Engineering = "engineering",
  ExoticWeapons = "exoticWeapons",
  Firearms = "firearms",
  Influence = "influence",
  Outdoors = "outdoors",
  Perception = "perception",
  Piloting = "piloting",
  Sorcery = "sorcery",
  Stealth = "stealth",
  Tasking = "tasking",
}